Authenticated updated

+++ app/Http/Controllers/ContactController.php
@@ -6,9 +6,16 @@ use App\Http\Resources\ContactDetailResource;
 use App\Http\Resources\ContactResource;
 use App\Models\Contact;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ContactController extends Controller
 {
+    // response properties
+    // success [ true , false]
+    // message
+    // errors
+
+
     /**
      * Display a listing of the resource.
      */
@@ -36,6 +43,7 @@ class ContactController extends Controller
             "name" => $request->name,
             "country_code" => $request->country_code,
             "phone_number" => $request->phone_number,
+            "user_id" => Auth::id(),
         ]);
         return new ContactDetailResource($contact);
     }
@@ -46,6 +54,14 @@ class ContactController extends Controller
     public function show(string $id)
     {
         $contact = Contact::find($id);
+        if (is_null($contact)) {
+            return response()->json([
+                // "success" => false,
+                "message" => "Contact not found",
+
+            ], 404);
+        }
+
 
         // SHOW DATA USING REGULAR WAY
         // return response()->json([
@@ -62,18 +78,42 @@ class ContactController extends Controller
     public function update(Request $request, string $id)
     {
         $request->validate([
-            "name" => "required",
-            "country_code" => "required|min:1|max:193",
-            "phone_number" => "required",
+            "name" => "nullable|min:1|max:20",
+            "country_code" => "nullable|integer|min:1|max:193",
+            "phone_number" => "nullable|min:7|max:15",
         ]);
 
         $contact = Contact::find($id);
 
-        $contact->update([
-            "name" => $request->name,
-            "country_code" => $request->country_code,
-            "phone_number" => $request->phone_number,
-        ]);
+        if (is_null($contact)) {
+            return response()->json([
+                // "success" => false,
+                "message" => "Contact not found",
+
+            ], 404);
+        }
+
+        // $contact->update([
+        //     "name" => $request->name,
+        //     "country_code" => $request->country_code,
+        //     "phone_number" => $request->phone_number,
+        // ]);
+
+        // $contact->update($request->all());
+
+        if ($request->has('name')) {
+            $contact->name = $request->name;
+        };
+
+        if ($request->has('country_code')) {
+            $contact->country_code = $request->country_code;
+        };
+
+        if ($request->has('phone_number')) {
+            $contact->phone_number = $request->phone_number;
+        };
+
+        $contact->update();
         return new ContactDetailResource($contact);
     }
 
@@ -83,8 +123,19 @@ class ContactController extends Controller
     public function destroy(string $id)
     {
         $contact = Contact::find($id);
+
+        if (is_null($contact)) {
+            return response()->json([
+                // "success" => false,
+                "message" => "Contact not found",
+
+            ], 404);
+        }
         $contact->delete();
 
-        return response()->json([],204);
+        // return response()->json([], 204);
+        return response()->json([
+            "message" => "contact deleted"
+        ]);
     }
 }
